resolve the enum value type

+++ src/CADL.Extension/Emitter.Csharp/src/lib/model.ts
@@ -364,7 +364,7 @@ export function getInputType(
 
         function enumMemberType(member: EnumMember): string {
             if (typeof member.value === "number") {
-                return "Int32";
+                return "Float64";
             }
             return "String";
         }
